+++ Packages/com.unity.mediation/Editor/CocoaPods~/UnityMediationAdColonyAdapter/UnityMediationAdColonyAdapter.framework/Headers/UMSAdColonyAdapterVersion.h
@@ -1,4 +1,0 @@
-#import <Foundation/Foundation.h>
-
-// Project version string for AdColonyAdapter.
-static NSString *const UMSAdColonyAdapterVersionString = @"0.0.11";