+++ Packages/com.unity.mediation/Editor/CocoaPods~/UnityMediationAdmobAdapter/UnityMediationAdmobAdapter.framework/Headers/UMSAdmobRewardedAdapter.h
@@ -1,8 +1,0 @@
-#import <Foundation/Foundation.h>
-#import <UnityMediationSdk/UnityMediationSdk.h>
-
-@interface UMSAdmobRewardedAdapter : NSObject <UMSRewardedAdapter>
-- (NSString *)               getVersion;
-
-- (id <UMSRewardedAdapterAd>)createRewardedAd:(UMSAdapterConfiguration *)parameters;
-@end