+++ Packages/com.unity.mediation/Editor/CocoaPods~/UnityMediationAppLovinAdapter/UnityMediationAppLovinAdapter.framework/Headers/UMSAppLovinAdapterVersion.h
@@ -1,4 +1,0 @@
-#import <Foundation/Foundation.h>
-
-// Project version string for AppLovinAdapter.
-static NSString *const UMSAppLovinAdapterVersionString = @"0.0.11";