+++ Packages/com.unity.mediation/Editor/CocoaPods~/UnityMediationFacebookAdapter/UnityMediationFacebookAdapter.framework/Headers/UMSFacebookAdapterVersion.h
@@ -1,4 +1,0 @@
-#import <Foundation/Foundation.h>
-
-// Project version string for FacebookAdapter.
-static NSString *const UMSFacebookAdapterVersionString = @"0.0.11";