+++ Packages/com.unity.mediation/Editor/CocoaPods~/UnityMediationIronSourceAdapter/UnityMediationIronSourceAdapter.framework/Headers/UMSIronSourceAdapterVersion.h
@@ -1,4 +1,0 @@
-#import <Foundation/Foundation.h>
-
-// Project version string for IronSourceAdapter.
-static NSString *const UMSIronSourceAdapterVersionString = @"0.0.11";