+++ Packages/com.unity.mediation/Editor/CocoaPods~/UnityMediationSdk/UnityMediationSdk.framework/Headers/UMSInitializationError.h
@@ -1,8 +1,0 @@
-#import <Foundation/Foundation.h>
-
-/**
- * Enum containing initialization error states.
- */
-typedef NS_ENUM (NSInteger, UMSInitializationError) {
-    UMSInitializationErrorUnknown
-};