+++ Packages/com.unity.mediation/Tests/Runtime/TestConstants.cs
@@ -3,6 +3,7 @@ namespace Unity.Mediation.Tests
     public class TestConstants
     {
         public const int Timeout = 10000;
+        public const string InvalidGameId = "InvalidGameId";
 
 #if UNITY_ANDROID
         public const string GameId = "3804049";
@@ -16,8 +17,8 @@ namespace Unity.Mediation.Tests
         public const string AdUnitInterstitialId = "interstitial_all_adunit";
         public const string AdUnitRewardedId = "rewarded_unity_adunit_1";
 #elif UNITY_IOS
-        public const string AdUnitInterstitialId = "interstitial_all_adunit";
-        public const string AdUnitRewardedId = "rewarded_all_adunit";
+        public const string AdUnitInterstitialId = "interstitial_am_adunit";
+        public const string AdUnitRewardedId = "rewarded_am_adunit";
 #else
         public const string AdUnitInterstitialId = "test-interstitial-id";
         public const string AdUnitRewardedId = "test-rewarded-id";